Move PropertyTreePath to another class, and fix regex
