+++ core/shared/src/main/scala/zio/config/package.scala
@@ -1,6 +1,9 @@
 package zio
 
 package object config extends KeyConversionFunctions with ConfigStringModule with ImplicitTupleConversion {
+  implicit class Interpolator(val sc: StringContext) extends AnyVal {
+    def path(str: String*): PropertyTreePath[String] = PropertyTreePath.$(sc.s(str :_*))
+  }
 
   implicit class MapOps[A](a: => A) {
     def toMap(config: ConfigDescriptor[A], keyDelimiter: String = "."): Either[String, Map[String, ::[String]]] =
